Provide for a signature which is an HMAC over the whole content
We were using a signature as a plain hash over most but not all the
content. For greater security we allow a signature made using
Digest::SHA's HMAC support, and calculated over the whole query content.

Thes signatures can only come from a client that supports them. We
therefore support the older mechanism as well for now. This needs to be
in place before we start propagating the required client change.