Attempt to fix bootstrap stylesheet issues
The bootstrap stylesheet we need for the accordian gadget is very
invasive and sets all sorts of things we don;t want set. This has caused
a complaint from TGL. Here we try to repair it by reverting some of
them that affect us the most, before including our own stylesheet.

Along the way, don't include the bootstrap stuff unless we need it.